Split apart experiment and measurement to break import cycle.
To eliminate a troublesome import cycle between query and composed,
this breaks those two into separate parts.  query now defines the subset
of Experiment that it uses in the form of a Protocol, to avoid depending
directly on experiment.py.
